+++ client-extensions/current-tickets-custom-element/src/main.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import {createRoot} from 'react-dom/client';
-import App from './App.jsx';
+import TicketApp from './TicketApp.jsx';
 import {QueryClient, QueryClientProvider} from 'react-query';
 
 const queryClient = new QueryClient();
@@ -10,7 +10,7 @@ class WebComponent extends HTMLElement {
 		root.render(
 			<React.StrictMode>
 				<QueryClientProvider client={queryClient}>
-					<App queryClient={queryClient} route={this.getAttribute('route') || '/'} />
+					<TicketApp queryClient={queryClient} route={this.getAttribute('route') || '/'} />
 				</QueryClientProvider>
 			</React.StrictMode>
 		);
